common: Support setting log levels at compile time
In order to support changing log levels per domain on systems where
getenv() is not available, now we allow passing SOL_LOG_LEVEL and
SOL_LOG_LEVELS as compile time definitions.
They work the same way as the environment variables on Unix systems,
with the only caveat that SOL_LOG_LEVEL must always be defined as a
string.
