+++ src/lib/common/sol-log-impl-riot.c
@@ -82,12 +82,6 @@ sol_log_impl_unlock(void)
 }
 
 void
-sol_log_impl_domain_init_level(struct sol_log_domain *domain)
-{
-    domain->level = _global_domain.level;
-}
-
-void
 sol_log_impl_print_function_stderr(void *data, const struct sol_log_domain *domain, uint8_t message_level, const char *file, const char *function, int line, const char *format, va_list args)
 {
     const char *name = domain->name ? domain->name : "";
