+++ src/lib/common/sol-log.c
@@ -38,6 +38,7 @@
 #include "sol-log.h"
 #include "sol-macros.h"
 #include "sol-log-impl.h"
+#include "sol-str-table.h"
 
 /* NOTE: these are not static since we share them with sol-log-impl-*.c
  * to make it simpler for those to change and check these variables
@@ -63,6 +64,8 @@ bool _show_line = true;
 void (*_print_function)(void *data, const struct sol_log_domain *domain, uint8_t message_level, const char *file, const char *function, int line, const char *format, va_list args) = sol_log_print_function_stderr;
 const void *_print_function_data = NULL;
 
+static struct sol_str_table *_env_levels = NULL;
+static char *_env_levels_str = NULL;
 static bool _inited = false;
 #define SOL_LOG_INIT_CHECK(fmt, ...) \
     do { \
@@ -95,18 +98,200 @@ void
 sol_log_shutdown(void)
 {
     sol_log_impl_shutdown();
+    free(_env_levels);
+    free(_env_levels_str);
+    _env_levels = NULL;
+    _env_levels_str = NULL;
     _inited = false;
+}
+
+static bool
+_level_str_parse(const char *buf, size_t buflen, uint8_t *storage)
+{
+    int16_t found;
+    static const struct sol_str_table table[] = {
+        SOL_STR_TABLE_ITEM("CRI",      SOL_LOG_LEVEL_CRITICAL),
+        SOL_STR_TABLE_ITEM("CRIT",     SOL_LOG_LEVEL_CRITICAL),
+        SOL_STR_TABLE_ITEM("CRITICAL", SOL_LOG_LEVEL_CRITICAL),
+        SOL_STR_TABLE_ITEM("DBG",      SOL_LOG_LEVEL_DEBUG),
+        SOL_STR_TABLE_ITEM("DEBUG",    SOL_LOG_LEVEL_DEBUG),
+        SOL_STR_TABLE_ITEM("ERR",      SOL_LOG_LEVEL_ERROR),
+        SOL_STR_TABLE_ITEM("ERROR",    SOL_LOG_LEVEL_ERROR),
+        SOL_STR_TABLE_ITEM("INF",      SOL_LOG_LEVEL_INFO),
+        SOL_STR_TABLE_ITEM("INFO",     SOL_LOG_LEVEL_INFO),
+        SOL_STR_TABLE_ITEM("WARN",     SOL_LOG_LEVEL_WARNING),
+        SOL_STR_TABLE_ITEM("WARNING",  SOL_LOG_LEVEL_WARNING),
+        SOL_STR_TABLE_ITEM("WRN",      SOL_LOG_LEVEL_WARNING),
+        { }
+    };
+
+    if (!sol_str_table_lookup(table, SOL_STR_SLICE_STR(buf, buflen), &found))
+        return false;
+    *storage = found;
+    return true;
+}
+
+static bool
+_int32_parse(const char *str, size_t size, int32_t *storage)
+{
+    int64_t value = 0;
+    size_t i;
+    bool negative = false;
+
+    errno = 0;
+    for (i = 0; i < size; i++) {
+        const char c = str[i];
+        if (i == 0 && c == '-') {
+            negative = true;
+            continue;
+        }
+
+        if (c < '0' || c > '9')
+            break;
+        value = (value * 10) + (c - '0');
+        if (value > INT32_MAX) {
+            errno = ERANGE;
+            return false;
+        }
+    }
+    if (i == 0) {
+        errno = EINVAL;
+        return false;
+    }
+    if (negative) {
+        value = -value;
+        if (value < INT32_MIN) {
+            errno = ERANGE;
+            return false;
+        }
+    }
+    *storage = value;
+    return true;
+}
+
+bool
+sol_log_level_parse(const char *str, size_t size, uint8_t *storage)
+{
+    int32_t i;
+    uint8_t level = 0;
+
+    if (size == 0)
+        return false;
+
+    if (_int32_parse(str, size, &i))
+        level = i;
+    else if (!_level_str_parse(str, size, &level))
+        return false;
+
+    *storage = level;
+    return true;
+}
+
+bool
+sol_log_levels_parse(const char *str, size_t size)
+{
+    const char *str_end = str + size;
+    unsigned int count = 0;
+    struct sol_str_table *itr;
+    const char *p, *e;
+
+    for (p = str; p < str_end; p++) {
+        if (*p == ',')
+            count++;
+    }
+    if (p > str)
+        count++;
+
+    free(_env_levels);
+    free(_env_levels_str);
+    if (count == 0) {
+        _env_levels = NULL;
+        _env_levels_str = NULL;
+        return false;
+    }
+
+    _env_levels = malloc((count + 1) * sizeof(*_env_levels));
+    if (!_env_levels) {
+        _env_levels_str = NULL;
+        return false;
+    }
+
+    _env_levels_str = strndup(str, size);
+    if (!_env_levels_str) {
+        free(_env_levels);
+        _env_levels = NULL;
+        _env_levels_str = NULL;
+        return false;
+    }
+
+    itr = _env_levels;
+    e = NULL;
+    str = _env_levels_str;
+    str_end = str + size;
+    for (p = str;; p++) {
+        if (p < str_end && *p == ':')
+            e = p;
+        else if (p == str_end || *p == ',') {
+            if (e && str + 1 < e && e + 1 < p) {
+                uint8_t val;
+                itr->key = str;
+                itr->len = e - str;
+                if (sol_log_level_parse(e + 1, p - e - 1, &val)) {
+                    itr->val = val;
+                    itr++;
+                }
+            }
+
+            if (p == str_end)
+                break;
+            str = p + 1;
+            e = NULL;
+        }
+    }
+
+    if (itr == _env_levels) {
+        free(_env_levels);
+        free(_env_levels_str);
+        _env_levels = NULL;
+        _env_levels_str = NULL;
+        return false;
+    }
+
+    itr->key = NULL;
+    itr->len = 0;
+    itr->val = 0;
+    return true;
+}
+
+SOL_API void
+sol_log_init_level_global(const char *str, size_t length)
+{
+    sol_log_level_parse(str, length, &_global_domain.level);
+}
+
+SOL_API void
+sol_log_init_levels(const char *str, size_t length)
+{
+    sol_log_levels_parse(str, length);
 }
 
 SOL_API void
 sol_log_domain_init_level(struct sol_log_domain *domain)
 {
+    int16_t level;
     SOL_LOG_INIT_CHECK("domain=%p", domain);
 
     if (!domain || !domain->name)
         return;
 
-    sol_log_impl_domain_init_level(domain);
+    if (_env_levels) {
+        level = sol_str_table_lookup_fallback(_env_levels,
+            sol_str_slice_from_str(domain->name),
+            _global_domain.level);
+    } else
+        level = _global_domain.level;
+
+    domain->level = level;
 }
 
 SOL_API void
